+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to
   ([#932]).
 - @cosmjs/stargate: Merge `DeliverTxFailure` and `DeliverTxSuccess` into a
   single `DeliverTxResponse` ([#878], [#949]). Add `assertIsDeliverTxFailure`.
+- @cosmjs/amino: Added `StdTx`, `isStdTx` and `makeStdTx` and removed them from
+  @cosmjs/launchpad. They are re-exported in @cosmjs/launchpad for backwards
+  compatibility.
 
 [#938]: https://github.com/cosmos/cosmjs/issues/938
 [#932]: https://github.com/cosmos/cosmjs/issues/932